requirements.txt
PyPDF2==3.0.1
reportlab==4.0.4
qrcode==7.4.2
Pillow==10.0.1 